riscv: Implement sub-word atomic exchange
The RISC-V ISA does not support sub-word atomic access.
Let's workaround this by preparing a shifted mask and
a shifted value, and replace the target bits within
a constrained LR/SC loop.

Based on the guarantees we get from the RISC-V specification,
constrained LR/SC loops provide the required forward progress
guarantee, so we won't run into a life-lock.

There have been similar approaches to solve this problem
before, but this approach minimizes the amount of instructions
within the read-modify-write loop (to an AND and an OR instruction).

Signed-off-by: Christoph Muellner <cmuellner@linux.com>